Refactor args/helpers in folders
