+++ inseq/commands/commands_utils/attribute_context/attribute_context_args.py
@@ -1,0 +1,213 @@
+from dataclasses import dataclass
+from typing import Any, Dict, List, Optional
+
+from .... import list_step_functions
+from ....attr.step_functions import is_contrastive_step_function
+from ....utils import cli_arg, pretty_dict
+from ..attribute.attribute_args import AttributeBaseArgs
+
+
+@dataclass
+class AttributeContextInputArgs:
+    input_current_text: str = cli_arg(
+        default="",
+        help=(
+            "The input text used for generation. If the model is a decoder-only model, the input text is a "
+            "prompt used for language modeling. If the model is an encoder-decoder model, the input text is the "
+            "source text provided as input to the encoder. It will be formatted as {current} in the "
+            "``input_template``."
+        ),
+    )
+    input_context_text: Optional[str] = cli_arg(
+        default=None,
+        help=(
+            "Additional input context influencing the generation of ``output_current_text``. If the model is a"
+            " decoder-only model, the input text is a prefix to the ``input_current_text`` prompt. If the model is an"
+            " encoder-decoder model, the input context is part of the source text provided as input to the encoder. "
+            " It will be formatted as {context} in the ``input_template``."
+        ),
+    )
+    input_template: Optional[str] = cli_arg(
+        default=None,
+        help=(
+            "The template used to format model inputs. The template must contain at least the"
+            " ``{current}`` placeholder, which will be replaced by ``input_current_text``. If ``{context}`` is"
+            " also specified, input-side context will be used. Can be modified for models requiring special tokens or"
+            " formatting in the input text (e.g. <brk> tags to separate context and current inputs)."
+            " Defaults to '{context} {current}' if ``input_context_text`` is provided, '{current}' otherwise."
+        ),
+    )
+    output_context_text: Optional[str] = cli_arg(
+        default=None,
+        help=(
+            "An output contexts for which context sensitivity should be detected. For encoder-decoder models, this"
+            " is a target-side prefix to the output_current_text used as input to the decoder. For decoder-only "
+            " models this is a portion of the model generation that should be considered as an additional context "
+            " (e.g. a chain-of-thought sequence). It will be formatted as {context} in the ``output_template``."
+            " If not provided but specified in the ``output_template``, the output context will be generated"
+            " along with the output current text, and user validation might be required to separate the two."
+        ),
+    )
+    output_current_text: Optional[str] = cli_arg(
+        default=None,
+        help=(
+            "The output text generated by the model when all available contexts are provided. Tokens in "
+            " ``output_current_text`` will be tested for context-sensitivity, and their generation will be attributed "
+            " to input/target contexts (if present) in case they are found to be context-sensitive. If specified, "
+            " this output is force-decoded. Otherwise, it is generated by the model using infilled ``input_template`` "
+            " and ``output_template``. It will be formatted as {current} in the ``output_template``."
+        ),
+    )
+    output_template: Optional[str] = cli_arg(
+        default=None,
+        help=(
+            "The template used to format model outputs. The template must contain at least the"
+            " ``{current}`` placeholder, which will be replaced by ``output_current_text``. If ``{context}`` is"
+            " also specified, output-side context will be used. Can be modified for models requiring special tokens or"
+            " formatting in the output text (e.g. <brk> tags to separate context and current outputs)."
+            " Defaults to '{context} {current}' if ``output_context_text`` is provided, '{current}' otherwise."
+        ),
+    )
+    has_input_context: bool = cli_arg(init=False)
+    has_output_context: bool = cli_arg(init=False)
+
+    def __post_init__(self):
+        if self.input_template is None:
+            self.input_template = "{current}" if self.input_context_text is None else "{context} {current}"
+        if self.output_template is None:
+            self.output_template = "{current}" if self.output_context_text is None else "{context} {current}"
+        self.has_input_context = "{context}" in self.input_template
+        self.has_output_context = "{context}" in self.output_template
+        if not self.input_current_text:
+            raise ValueError("--input_current_text must be a non-empty string.")
+        if self.input_context_text and not self.has_input_context:
+            logger.warning(
+                f"input_template has format {self.input_template} (no {{context}}), but --input_context_text is"
+                " specified. Ignoring provided --input_context_text."
+            )
+            self.input_context_text = None
+        if self.output_context_text and not self.has_output_context:
+            logger.warning(
+                f"output_template has format {self.output_template} (no {{context}}), but --output_context_text is"
+                " specified. Ignoring provided --output_context_text."
+            )
+            self.output_context_text = None
+        if not self.input_context_text and self.has_input_context:
+            raise ValueError(
+                f"{{context}} format placeholder is present in input_template {self.input_template},"
+                " but --input_context_text is not specified."
+            )
+        if "{current}" not in self.input_template:
+            raise ValueError(f"{{current}} format placeholder is missing from input_template {self.input_template}.")
+        if "{current}" not in self.output_template:
+            raise ValueError(f"{{current}} format placeholder is missing from output_template {self.output_template}.")
+        if not self.input_current_text:
+            raise ValueError("--input_current_text must be a non-empty string.")
+        if self.has_input_context and self.input_template.find("{context}") > self.input_template.find("{current}"):
+            raise ValueError(
+                f"{{context}} placeholder must appear before {{current}} in input_template '{self.input_template}'."
+            )
+        if self.has_output_context and self.output_template.find("{context}") > self.output_template.find("{current}"):
+            raise ValueError(
+                f"{{context}} placeholder must appear before {{current}} in output_template '{self.output_template}'."
+            )
+        if not self.output_template.endswith("{current}"):
+            *_, suffix = self.output_template.partition("{current}")
+            logger.warning(
+                f"Suffix '{suffix}' was specified in output_template and will be used to ignore the specified suffix"
+                " tokens during context sensitivity detection. Make sure that the suffix corresponds to the end of the"
+                " output_current_text by forcing --output_current_text if necessary."
+            )
+
+
+@dataclass
+class AttributeContextMethodArgs(AttributeBaseArgs):
+    context_sensitivity_metric: str = cli_arg(
+        default="kl_divergence",
+        help="The contrastive metric used to detect context-sensitive tokens in ``output_current_text``.",
+        choices=[fn for fn in list_step_functions() if is_contrastive_step_function(fn)],
+    )
+    align_output_context_auto: bool = cli_arg(
+        default=False,
+        help=(
+            "Argument used for encoder-decoder model when generating text with an output template including both"
+            " {context} and {current}, to attempt an automatic detection of which parts of the output belong to"
+            " context vs. current in absence of other explicit cues. If set to True, the input and output context"
+            " and current texts are aligned automatically (assuming an MT-like task), and the alignments are "
+            " assumed to be valid to separate the two without further user validation. Otherwise, the user is "
+            " prompted to manually specify which part of the generated text corresponds to the output context."
+        ),
+    )
+    special_tokens_to_keep: List[str] = cli_arg(
+        default_factory=list,
+        help="Special tokens to preserve in the generated string, e.g. ``<brk>`` separator between context and current.",
+    )
+    context_sensitivity_std_threshold: float = cli_arg(
+        default=1.0,
+        help=(
+            "Parameter to control the selection of ``output_current_text`` tokens considered as context-sensitive for "
+            "moving onwards with attribution. Corresponds to the number of standard deviations above or below the mean"
+            " ``context_sensitivity_metric`` score for tokens to be considered context-sensitive."
+        ),
+    )
+    context_sensitivity_topk: Optional[int] = cli_arg(
+        default=None,
+        help=(
+            "If set, after selecting the salient context-sensitive tokens with ``context_sensitivity_std_threshold`` "
+            "only the top-K remaining tokens are used. By default no top-k selection is performed."
+        ),
+    )
+    attribution_std_threshold: float = cli_arg(
+        default=1.0,
+        help=(
+            "Parameter to control the selection of ``input_context_text`` and ``output_context_text`` tokens "
+            "considered as salient as a result for the attribution process. Corresponds to the number of standard "
+            "deviations above or below the mean ``attribution_method`` score for tokens to be considered salient. "
+            "CCI scores for all context tokens are saved in the output, but this parameter controls which tokens are "
+            "used in the visualization of context reliance."
+        ),
+    )
+    attribution_topk: Optional[int] = cli_arg(
+        default=None,
+        help=(
+            "If set, after selecting the most salient tokens with ``attribution_std_threshold`` "
+            "only the top-K remaining tokens are used. By default no top-k selection is performed."
+        ),
+    )
+
+
+@dataclass
+class AttributeContextOutputArgs:
+    show_intermediate_outputs: bool = cli_arg(
+        default=False,
+        help=(
+            "If specified, the intermediate outputs produced by the Inseq library for context-sensitive target "
+            "identification (CTI) and contextual cues imputation (CCI) are shown during the process.",
+        ),
+    )
+    save_path: Optional[str] = cli_arg(
+        default=None,
+        aliases=["-o"],
+        help="If present, the output of the two-step process will be saved in JSON format at the specified path.",
+    )
+    add_output_info: bool = cli_arg(
+        default=True,
+        help="If specified, additional information about the attribution process is added to the saved output.",
+    )
+    viz_path: Optional[str] = cli_arg(
+        default=None,
+        help="If specified, the visualization produced from the output is saved in HTML format at the specified path.",
+    )
+    show_viz: bool = cli_arg(
+        default=True,
+        help="If specified, the visualization produced from the output is shown in the terminal.",
+    )
+
+
+@dataclass
+class AttributeContextArgs(AttributeContextInputArgs, AttributeContextMethodArgs, AttributeContextOutputArgs):
+    def __repr__(self):
+        return f"{self.__class__.__name__}({pretty_dict(self.__dict__)})"
+
+    def to_dict(self) -> Dict[str, Any]:
+        return dict(self.__dict__.items())
